@frenchy64 2021-09-11
Thanks, there's a bunch of redundancy and extra allocation in the superclass logic, I will look at that next.

Pretty much the same on my machine: https://github.com/frenchy64/fully-satisfies/commit/6fdb077c0232aed7d70fb7586f46e8157d9d70c5