@wandersoncferreira 2021-09-12
Interesting that your benchmarks results are different from mine. I just run `lein bench` with the code in master and got

```clojure
bench satisfies?
Evaluation count : 4911060 in 60 samples of 81851 calls.
             Execution time mean : 12.381497 µs
    Execution time std-deviation : 164.923412 ns
   Execution time lower quantile : 12.179057 µs ( 2.5%)
   Execution time upper quantile : 12.720163 µs (97.5%)
                   Overhead used : 6.876313 ns

Found 3 outliers in 60 samples (5.0000 %)
	low-severe	 2 (3.3333 %)
	low-mild	 1 (1.6667 %)
 Variance from outliers : 1.6389 % Variance is slightly inflated by outliers


bench fully-satisfies?
Evaluation count : 61687140 in 60 samples of 1028119 calls.
             Execution time mean : 966.467816 ns
    Execution time std-deviation : 5.440150 ns
   Execution time lower quantile : 962.115441 ns ( 2.5%)
   Execution time upper quantile : 974.652734 ns (97.5%)
                   Overhead used : 6.876313 ns

Found 5 outliers in 60 samples (8.3333 %)
	low-severe	 3 (5.0000 %)
	low-mild	 2 (3.3333 %)
 Variance from outliers : 1.6389 % Variance is slightly inflated by outliers
```

but your results both versions are in the microseconds unit.