@frenchy64 2021-09-12
We probably need to compare their relative performance for consistent numbers.